# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OneStyle API",
    "description": "Express + PostgreSQL backend for OneStyle e-commerce app",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Local dev" },
    { "url": "https://your-render-backend.onrender.com", "description": "Production (Render)" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "required": ["id", "name", "category", "image_url", "new_price", "old_price"],
        "properties": {
          "id":                { "type": "integer", "example": 1 },
          "name":              { "type": "string",  "example": "Casual Striped Blouse with Peplum Hem" },
          "category":          { "type": "string",  "enum": ["women", "men", "kid"] },
          "image_url":         { "type": "string",  "example": "product_1.png" },
          "new_price":         { "type": "number",  "format": "float", "example": 50.0 },
          "old_price":         { "type": "number",  "format": "float", "example": 80.5 },
          "description":       { "type": "string",  "nullable": true, "example": "A stylish striped blouse..." },
          "is_new_collection": { "type": "boolean", "example": false },
          "is_popular":        { "type": "boolean", "example": true },
          "created_at":        { "type": "string",  "format": "date-time" }
        }
      },
...
      "Review": {
        "type": "object",
        "required": ["id", "product_id", "user_id", "rating", "created_at"],
        "properties": {
          "id":            { "type": "integer", "example": 1 },
          "product_id":    { "type": "integer", "example": 3 },
          "user_id":       { "type": "integer", "example": 7 },
          "reviewer_name": { "type": "string",  "example": "Jane Doe" },
          "rating":        { "type": "integer", "minimum": 1, "maximum": 5, "example": 4 },
          "comment":       { "type": "string",  "nullable": true, "example": "Great quality!" },
          "created_at":    { "type": "string",  "format": "date-time" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string", "example": "Something went wrong." }
        }
      },
      "MessageResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string", "example": "Success." }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/products": {
      "get": {
        "operationId": "getProducts",
        "summary": "List products with optional filtering and sorting",
        "tags": ["Products"],
        "parameters": [
          { "name": "category",   "in": "query", "schema": { "type": "string", "enum": ["women", "men", "kid"] }, "description": "Filter by category" },
          { "name": "search",     "in": "query", "schema": { "type": "string" }, "description": "Full-text search on name and description" },
          { "name": "priceRange", "in": "query", "schema": { "type": "string", "enum": ["low", "medium", "high"] }, "description": "low = under 100, medium = 100–200, high = over 200" },
          { "name": "sort",       "in": "query", "schema": { "type": "string", "enum": ["price-asc", "price-desc", "name-asc"] }, "description": "Sort order" }
        ],
        "responses": {
          "200": {
            "description": "Product list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["products"],
                  "properties": {
                    "products": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid filter value", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Internal server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/products/new-collections": {
      "get": {
        "operationId": "getNewCollections",
        "summary": "Get products flagged as new collection",
        "tags": ["Products"],
        "responses": {
          "200": {
            "description": "New collection products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "products": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } } }
                }
              }
            }
          },
          "500": { "description": "Internal server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/products/popular": {
      "get": {
        "operationId": "getPopularProducts",
        "summary": "Get products flagged as popular",
        "tags": ["Products"],
        "responses": {
          "200": {
            "description": "Popular products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "products": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } } }
                }
              }
            }
          },
          "500": { "description": "Internal server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "operationId": "getProductById",
        "summary": "Get a single product by ID",
        "tags": ["Products"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 }
        ],
        "responses": {
          "200": {
            "description": "Product detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["product"],
                  "properties": { "product": { "$ref": "#/components/schemas/Product" } }
                }
              }
            }
          },
          "400": { "description": "Invalid id", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Product not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Internal server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/products/{id}/reviews": {
      "get": {
        "operationId": "getProductReviews",
        "summary": "Get reviews and aggregate rating for a product",
        "tags": ["Reviews"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 }
        ],
        "responses": {
          "200": {
            "description": "Reviews and aggregate",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["reviews", "average_rating", "total_reviews"],
                  "properties": {
                    "reviews":        { "type": "array", "items": { "$ref": "#/components/schemas/Review" } },
                    "average_rating": { "type": "number", "format": "float", "example": 4.2 },
                    "total_reviews":  { "type": "integer", "example": 15 }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid id", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Internal server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/reviews": {
      "post": {
        "operationId": "submitReview",
        "summary": "Submit or update a product review (one per user per product)",
        "tags": ["Reviews"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["productId", "rating"],
                "properties": {
                  "productId": { "type": "integer", "example": 3 },
                  "rating":    { "type": "integer", "minimum": 1, "maximum": 5, "example": 4 },
                  "comment":   { "type": "string",  "example": "Great fit and quality!" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["review"],
                  "properties": { "review": { "$ref": "#/components/schemas/Review" } }
                }
              }
            }
          },
          "400": { "description": "Invalid productId or rating", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Product not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Internal server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/reviews/{reviewId}": {
      "delete": {
        "operationId": "deleteReview",
        "summary": "Delete the authenticated user's own review",
        "tags": ["Reviews"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "reviewId", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 7 }
        ],
        "responses": {
          "200": { "description": "Review deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageResponse" } } } },
          "400": { "description": "Invalid reviewId", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Review not found or not owned by you", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Internal server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    }
  }
}